+++ DynamicCavemanWorld/Assets/Classes/TempEquation.cs
@@ -4,13 +4,13 @@ using System.Collections;
 public class TempEquation {
 
     // Variables
-    private decimal A;
-    private decimal B;
-    private decimal C;
-    private decimal D;
-    private decimal E;
-    private decimal F;
-    private decimal G;
+    private double A;
+    private double B;
+    private double C;
+    private double D;
+    private double E;
+    private double F;
+    private double G;
 
     // Constructor: takes in the input from the variable semi-static temperature references
     public TempEquation()
